requirements.txt
black==22.3.0
click==8.1.2
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
tomli==2.0.1
Werkzeug==2.1.1
